requirements.txt
python-telegram-bot
Pillow
numpy
dlib